Moved NamedAccounts into Cluster.
