+++ tests/nodeos_under_min_avail_ram.py
@@ -2,6 +2,7 @@
 
 from core_symbol import CORE_SYMBOL
 from Cluster import Cluster
+from Cluster import NamedAccounts
 from WalletMgr import WalletMgr
 from Node import Node
 from TestHelper import TestHelper
@@ -25,39 +26,6 @@ import re
 
 Print=Utils.Print
 errorExit=Utils.errorExit
-
-class NamedAccounts:
-
-    def __init__(self, cluster, numAccounts):
-        Print("NamedAccounts %d" % (numAccounts))
-        self.numAccounts=numAccounts
-        self.accounts=cluster.createAccountKeys(numAccounts)
-        if self.accounts is None:
-            errorExit("FAILURE - create keys")
-        accountNum = 0
-        for account in self.accounts:
-            Print("NamedAccounts Name for %d" % (accountNum))
-            account.name=self.setName(accountNum)
-            accountNum+=1
-
-    def setName(self, num):
-        retStr="test"
-        digits=[]
-        maxDigitVal=5
-        maxDigits=8
-        temp=num
-        while len(digits) < maxDigits:
-            digit=(num % maxDigitVal)+1
-            num=int(num/maxDigitVal)
-            digits.append(digit)
-
-        digits.reverse()
-        for digit in digits:
-            retStr=retStr+str(digit)
-
-        Print("NamedAccounts Name for %d is %s" % (temp, retStr))
-        return retStr
-
 
 args = TestHelper.parse_args({"--dump-error-details","--keep-logs","-v","--leave-running","--clean-run","--wallet-port"})
 Utils.Debug=args.v
